Auto service started

+++ src/main/java/com/example/Uncha_Muncha_Bot/dto/AutomobileServiceTypeDTO.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Getter
 public class AutomobileServiceTypeDTO {
     private Long id;
-    private Integer AutoServiceId;
+    private Long AutoServiceId;
     private String serviceName;
 }
